ponzgen/requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
supabase==2.14.0
# python-dotenv==1.0.0
python-dotenv
pydantic==2.10.5
httpx==0.28.1
langchain==0.3.14
langchain-openai==0.3.0
langchain-mcp-adapters==0.0.3
langgraph==0.2.62
langchain-mcp-adapters==0.0.3
#mcp-proxy==0.5.1
mcp-proxy==0.6.0
schedule
psutil==7.0.0
# aiohttp==3.11.14
aiohttp
fal-client
# fastmcp==2.2.1
fastmcp
langchain-community==0.3.14